# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing cats and feedback",
        "title": "Cat Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Get a random cat image from the external API and render it on the homepage",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Get random cat image",
                "responses": {
                    "200": {
                        "description": "HTML page with random cat image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cats": {
            "get": {
                "description": "Get list of all cats preview with optional limit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Get cats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of cats returned",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cat_service_internal_data_dto_cats.CatPreview"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new cat with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Create a new cat",
                "parameters": [
                    {
                        "description": "Cat information",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cat_service_internal_data_dto_cats.CatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/cat_service_internal_data_dto_cats.CatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cats/feedback": {
            "get": {
                "description": "Save feedback for a cat",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Save feedback",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quality rating (1-5)",
                        "name": "quality",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cuteness rating (1-5)",
                        "name": "cute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Additional feedback message",
                        "name": "message",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Feedback saved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cats/{id}": {
            "get": {
                "description": "Get detailed information about a cat by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Get cat by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cat_service_internal_data_dto_cats.CatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/views/cats/internal/{id}": {
            "get": {
                "description": "Get a rendered view of a cat by its ID",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Get cat view by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view of the cat",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cat_service_internal_data_dto_cats.CatPreview": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "cat_service_internal_data_dto_cats.CatRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "homeless": {
                    "type": "boolean"
                },
                "img_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "cat_service_internal_data_dto_cats.CatResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "homeless": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}